requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sentence-transformers==2.7.0
faiss-cpu==1.8.0
scikit-learn==1.5.2
sqlalchemy==2.0.35
greenlet==3.0.3
aiosqlite==0.20.0
asyncpg==0.29.0
httpx==0.27.2
numpy==1.26.4
pydantic==2.9.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
tenacity>=8.0.0
tqdm>=4.66.0
